// src/types/enums.ts
export enum MainPath {
  Contact = 'contact',
  Root = '/',
}

export enum LeftNavPath {
  AllProducts = 'allprodutcs',
  Electronics = 'electronics',
  Jewelery = 'jewelery',
  Men = 'men',
  Women = 'women',
}

export enum Title {
  AllProducts = 'All Products',
  Contact = 'Contact',
  Electronics = 'Electronics',
  Home = 'Home',
  Jewelery = 'Jewelery',
  MensClothing = "Men's clothing",
  WomensClothing = "Women's clothing",
}

export enum ProductApi {
  Electronics = 'electronics',
  Jewelery = 'jewelery',
  MensClothing = "men's clothing",
  WomensClothing = "women's clothing",
}

export enum KeyCode {
  Alt = 'AltLeft',
  ArrowDown = 'ArrowDown',
  ArrowLeft = 'ArrowLeft',
  ArrowRight = 'ArrowRight',
  ArrowUp = 'ArrowUp',
  End = 'End',
  Enter = 'Enter',
  Esc = 'Escape',
  Home = 'Home',
  Shift = 'ShiftLeft',
  Space = 'Space',
  Tab = 'Tab',
}

export enum IconName {
  Logo = 'logo',
}

export enum SizeVariant {
  Lg = 'lg',
  Md = 'md',
  Sm = 'sm',
}

export enum BtnVariant {
  Ghost = 'ghost',
  Link = 'link',
  Neutral = 'neutral',
  Primary = 'primary',
  Secondary = 'secondary',
}

export enum ToastType {
  Error = 'error',
  Info = 'info',
  Success = 'success',
  Warning = 'warning',
}

export enum ToastRole {
  Alert = 'alert',
  Status = 'status',
}